Move policy handler to new home
