+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/Errors/Converters/Authorization/GridOperatorIsNotOwnerOfMeteringPointErrorConverter.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.MeteringPoints.Application.Authorization.GridOperatorPolicies;
 using Energinet.DataHub.MeteringPoints.Application.ChangeMasterData.Consumption;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors.Converters.Authorization
